fix: toggle layer content by clicking the entire layer header
