+++ .changeset/shy-starfishes-divide.md
@@ -1,0 +1,5 @@
+---
+"geohub": patch
+---
+
+fix: toggle layer content by clicking the entire layer header
